게시판 (조건 만족 -> 글수정, 삭제)

+++ board/views.py
@@ -56,15 +56,21 @@ def modify(request):
 def modifyform(request):
     id = request.GET['id']
     board = Board.objects.get(id=id)
-    context = {'board': board}
-
-    return render(request, 'board/modify.html', context)
-
+    user_id = request.session['authuser']['id']
+    print(board.user_id)
+    print(user_id)
+    if board.user_id == user_id:
+        context = {'board': board}
+        return render(request, 'board/modify.html', context)
+    else:
+        return HttpResponseRedirect('/board')
 
 def delete(request):
     id = request.GET['id']
     board = Board.objects.get(id=id)
-    if board.name == request.session['authuser']['name']:
-        Board.objects.filter(id=id).delete()
+    print(board.user_id)
+    user_id = request.session['authuser']['id']
+    print(user_id)
+    Board.objects.filter(id=id).filter(user_id=user_id).delete()
 
     return HttpResponseRedirect('/board')
